Refactor for testing
The split here isn't as clean as it is with Luca's original
implementation using actix. I'm passing an axum::routing::Router back,
which incorporates the handling logic, but leaves the binding and Server
creation up to the caller. Axum's Server doesn't seem to lend itself to
be being passed back after having a Router attached, but I could also
just be daft.