Add email_client and refactor initialization
Postmark doesn't let schlubs with a GMail address sign-up, so we're
using MailChimp. Yay! This makes verifying the JSON payload a little
different and I'm sure that it can be refined.

Since we're refactoring for better modularity and testing, I revisited
where the split was (returning a Router instead of a Server).
`axum_server::Server` can't be returned after consuming the `Builder` by
calling `::serve`, but the underlying `hyper::Server` can! So, we're
that much closer to parity in design and it's certainly cleaner.

This isn't an atomic commit because I broke things part way through and
felt like I needed to fix them before committing. Bad, Developer, bad!