Improve error logging for subscription handler
This is about as close as I can get to parity with Luca's implementation
with actix_web. Axum doesn't want to acknowledge that things fail I
guess, so I'm using a call to the `tracing::error` macro in the
conversion of my handler error type to an `axum::response::Response` to
work around that. The http request span doesn't reiterate the error
message, though.