Improve error-handling in confirmation endpoint
Upon further examination, we have another regression compared to using
actix_web: requests missing query parameters or form data are handled by
the framework in both cases. In actix_web, that issues a trace event at
the warning level; but, with axum, it goes unlogged. Conceiveably, I
could make the fields optional and then handle it myself, but that's a
bunch of work to get something that I should already be getting!