First pass of authentication
This includes wrong implementation of flash errors (yay!) and also
finally figuring out how to extract substates using the
`axum::extract::State` extractor. This makes it so that handlers that
only need to connect to Postgres extract only the connection pool and
don't have unnecessary access to the email client or HMAC secret.